checkpoint antes da refatoraçao

+++ controle-treinos/app/Http/Controllers/API/API_ExerciseController.php
@@ -12,54 +12,90 @@ class API_ExerciseController extends Controller
     {
         $exercises = Exercise::all();
         
-        return response()->json($exercises);
+        return response()->json($this->getExercices());
+    }
+
+    public function getExercices()
+    {
+        return Exercise::all();
     }
 
     public function show(int $id)
     {
         $exercise = Exercise::find($id);
 
-        return response()->json($exercise);
+        if ($exercise) {
+            return response()->json($exercise);
+        } else {
+            return response()->json(
+                ["error" => "Exercise not found"]
+            );
+            
+        }
     }
     
     public function store(Request $request)
     {
-        $exercise = Exercise::create($request->all());
-        
-        return response()->json($exercise);
+        if (!$request->name) 
+        {
+            return response()->json(
+                ["error" => "Property 'name' not found"]
+            );
+        } else {
+            $exercise = Exercise::create($request->all());
+            
+            return response()->json($exercise);
+        }
     }
 
     public function update(Request $request, int $id)
     {
 
-        $exercise = Exercise::find($id)->update($request->all());
+        $exercise = Exercise::find($id);
+        
+        if ($exercise)
+        {
+            $exercise->update($request->all());
+        
+            return response()->json($exercise);
+        } else {
+            return response()->json(
+                ["error" => "Exercise not found"]
+            );
+        }
 
-        return response()->json($exercise);
     }
 
     public function destroy($id)
     {
         $exercise = Exercise::find($id);
-        $workouts = $exercise->workouts;
 
-        foreach($workouts as $workout)
-        {
-            if($workout->active == true)
+        if ($exercise) {
+            $workouts = $exercise->workouts;
+    
+            foreach($workouts as $workout)
             {
-                return response()->json(
-                    ["error" => "Exercício em treino ativo"]
-                );
+                if($workout->active == true)
+                {
+                    return response()->json(
+                        ["error" => "Exercício em treino ativo"]
+                    );
+                }
             }
+    
+            foreach($workouts as $workout)
+            {
+                $workout->exercises()->detach();
+            }
+    
+            $exercise = Exercise::destroy($id);
+    
+            return response()->json($exercise);
+        } else {
+            return response()->json(
+                ["error" => "Exercise not found"]
+            );
         }
-
-        foreach($workouts as $workout)
-        {
-            $workout->exercises()->detach();
-        }
-
-        $exercise = Exercise::destroy($id);
-
-        return response()->json($exercise);
     }
 
     public function showEx(int $id)
